adp536x: move from /drivers to /lib
Move from /drivers to /lib, since it does not implement device.h
