+++ lib/adp536x/CMakeLists.txt
