+++ lib/adp536x/adp536x.c
